requirements.txt
Django>=3.2,<4.0
djangorestframework>=3.13.1,<3.14
psycopg2>=2.8.6,<2.9
drf-spectacular>=0.15.1,<0.16
django-crispy-forms>=1.13.0
uwsgi>=2.0.20,<2.1
daphne==4.0.0
Pillow==9.3.0
six==1.16.0
crispy-bootstrap5==0.7
django-crispy-forms==2.0
django-crispy-forms
django-phonenumber-field==7.0.2
phonenumbers
Babel==2.12.1
django-cors-headers==3.13.0
django-recaptcha==3.0.0
djangorestframework-simplejwt==5.2.2
celery
redis==3.5.3
django-celery-beat
django_celery_results==2.2.0
django-cors-headers==3.13.0
google-api-python-client==1.7.2
google-auth==1.8.0
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
oauth2client
django-widget-tweaks
django-ckeditor==6.7.0
requests
certifi==2022.6.15
charset-normalizer==2.1.0
idna==3.3
requests==2.28.1
urllib3==1.26.9
djangorestframework-simplejwt
djangorestframework-simplejwt[token_blacklist]

